+++ yash-semantics/CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to `yash-semantics` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] - Unreleased
+## [0.3.0] - Unreleased
 
 ### Changed
 
 - External dependency versions:
     - Rust 1.75.0 → 1.77.0
+- The first argument to `runner::ReadEvalLoop::new` is now a `&RefCell<&mut Env>`
+  instead of a `&mut Env`. This allows `yash_syntax::input::Input` implementors
+  to access the environment as it reads input.
 
 ### Removed
 
@@ -73,6 +76,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of the `yash-semantics` crate
 
-[0.2.1]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.2.1
+[0.3.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.3.0
 [0.2.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.2.0
 [0.1.0]: https://github.com/magicant/yash-rs/releases/tag/yash-semantics-0.1.0